eventsmanager shouldnt be in core, just moving stuff so it compiles when u just exclude every1's folders
