+++ FrankenToilet/duviz/events/EventsCreator.cs
@@ -1,5 +1,4 @@
-﻿using FrankenToilet.Core;
-using System.Collections;
+﻿using System.Collections;
 using UnityEngine;
 
 namespace FrankenToilet.duviz.events;
